[FEATURE] content dependant notification e-mail subject
The subject of the notification E-mail contains now information about the touched main nodes, the backenduser and the stage.

The main node is the first node of the touched rootline. In some cases it might be helpful to exclude some nodes. This can be configured by adding a comma separated list of nodes in the notificationSubjectExcludedNodes property